+++ Local-Sales/app/assets/components/ListItem.js
@@ -34,7 +34,6 @@ const styles = StyleSheet.create({
         width: 100,
         height: 100,
         borderRadius: 50,
-        borderTopStartRadius:10,
         margin: 10,
     },
     title:{
