+++ Local-Sales/package.json
@@ -9,6 +9,7 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@expo/vector-icons": "^13.0.0",
     "@fortawesome/fontawesome-svg-core": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-native-fontawesome": "^0.3.0",
@@ -20,8 +21,9 @@
     "expo": "~49.0.8",
     "expo-constants": "~14.4.2",
     "expo-status-bar": "~1.6.0",
+    "formik": "^2.4.5",
     "react": "18.2.0",
-    "react-native": "0.72.5",
+    "react-native": "^0.72.5",
     "react-native-gesture-handler": "~2.12.0",
     "react-native-reanimated": "~3.3.0",
     "react-native-safe-area-context": "4.6.3",
